+++ synthesizer/src/vm/tests/mod.rs
@@ -26,4 +26,7 @@ mod test_v10;
 mod test_v11;
 
 #[cfg(feature = "test")]
+mod test_v13;
+
+#[cfg(feature = "test")]
 use super::*;
